Add files via upload
This project’s purpose is to evaluate and analyze the IMDb dataset to gain insight into the performance of performers, films, and their networks. The project comprises running SQL queries and Python code on the dataset to extract important information, as well as constructing a Tableau dashboard to present the results in an interactive and visually appealing manner.

IMDb is a website that contains information about movies, television shows, podcasts, home videos, video games, and streaming content online, such as cast, production staff, and personal biographies; narrative summaries; trivia; ratings; and fan and critical reviews. IMDb began as a fan-run movie database on the Usenet forum “rec.arts.movies” in 1990, then transferred to the Web in 1993. It has been owned and operated by IMDb.com, Inc., a subsidiary of Amazon, since 1998.

The database included 10.1 million titles (including television episodes) and 11.5 million human records as of March 2022. Furthermore, the website had 83 million registered users.